Actors were moved to separate package
