+++ src/main/scala/sk/drunkenpanda/bot/App.scala
@@ -1,10 +1,8 @@
 package sk.drunkenpanda.bot
 
-import akka.actor.Actor
-import akka.actor.ActorRef
 import akka.actor.ActorSystem
-import akka.actor.Props
 import java.net.Socket
+import sk.drunkenpanda.bot.actor._
 import sk.drunkenpanda.bot.io._
 import sk.drunkenpanda.bot.plugins.AbstractPluginModule
 import sk.drunkenpanda.bot.plugins.EchoPlugin
@@ -21,25 +19,13 @@ object App {
     val writer = system.actorOf(StreamWriter.props(bot, source))
     val reader = system.actorOf(StreamReader.props(bot, source))
     val processor = system.actorOf(MessageProcessor.props(new SimplePluginModule))
-    val masterActor = system.actorOf(Props(classOf[MasterActor], 
-      processor, reader, writer))
-    masterActor ! StreamReader.Start
+    val masterActor = system.actorOf(
+      MasterActor.props(processor, reader, writer))
+    masterActor ! MasterActor.Start
   }
   
   def main(args: Array[String]): Unit = startBot(source)
 
-  class MasterActor(processor: ActorRef, reader: ActorRef, writer: ActorRef) 
-    extends Actor {
-    
-    def receive = {
-      case StreamReader.Start => reader ! StreamReader.Start
-      case m: PrivateMessage => processor ! m
-      case m: Ping => processor ! m
-      case m: Pong => writer ! m
-      case m: Response => writer ! m
-    }
-
-}
  
   class SimplePluginModule extends AbstractPluginModule {
